requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
httpx>=0.24.0
tabulate>=0.9.0
pytest>=7.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

